requirements.txt
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pytesseract>=0.3.10
anthropic>=0.7.0
transformers>=4.30.0
scikit-image>=0.21.0
matplotlib>=3.7.0
jinja2>=3.1.0
